+++ .env
@@ -1,5 +1,13 @@
+
+# development
 # NEXT_PUBLIC_RELAY_URL=ws://localhost:8008
 # NEXT_PUBLIC_LNURL_PROTOCOL=http
+# # This value is the hex pubkey derived from the wavman-tools SK
+# NEXT_PUBLIC_TRACK_EVENT_PUBKEY=e0e07e122493db0355e8b298c6bbf7a9423212d8f10f3a06b5af92de9cab0131
+# NEXT_PUBLIC_ENABLE_RANDOM_TRACKS=false
+
+# production
 NEXT_PUBLIC_RELAY_URL=wss://relay.wavlake.com/
 NEXT_PUBLIC_LNURL_PROTOCOL=https
+NEXT_PUBLIC_TRACK_EVENT_PUBKEY=7759fb24cec56fc57550754ca8f6d2c60183da2537c8f38108fdf283b20a0e58
 NEXT_PUBLIC_ENABLE_RANDOM_TRACKS=false
