+++ src/nostr/zapUtils.ts
@@ -3,18 +3,14 @@ import {
   generatePrivateKey,
   getPublicKey,
   Event,
+  getEventHash,
+  signEvent,
 } from "nostr-tools";
 
 const protocol = process.env.NEXT_PUBLIC_LNURL_PROTOCOL;
 
 export const sats2millisats = (amount: number) => amount * 1000;
 export const chopDecimal = (amount: number) => Math.floor(amount);
-export const generateLNURLFromZapTag = (zapTag: string[]) => {
-  const [zap, zapAddress, lud] = zapTag;
-  const [username, domain] = zapAddress.split("@");
-  if (!username || !domain) return false;
-  return `${protocol}://${domain}/.well-known/lnurlp/${username}`;
-};
 export const validateNostrPubKey = (nostrPubKey: string) => {
   if (
     nostrPubKey == null ||
@@ -51,7 +47,7 @@ export const signZapEventNip07 = async ({
   recepientPubKey: string;
   zappedEvent: Event;
   pubkey: string;
-}): Promise<Event | void> => {
+}): Promise<Event> => {
   try {
     const unsignedEvent: UnsignedEvent = {
       kind: 9734,
@@ -70,8 +66,117 @@ export const signZapEventNip07 = async ({
     const signedEvent: Event | undefined = await window.nostr?.signEvent?.(
       unsignedEvent
     );
+    if (!signedEvent) {
+      // TODO implement a fall back route to sign anon
+      throw "No signed event";
+    }
     return signedEvent;
   } catch (err) {
-    console.log("error signing zap event", { err, lnurl, zappedEvent });
+    console.error({ err, lnurl, zappedEvent });
+    throw "Error signing event with NIP-07";
   }
+};
+
+export const signAnonZapEvent = async ({
+  content,
+  amount,
+  lnurl,
+  recepientPubKey,
+  zappedEvent,
+}: {
+  content: string;
+  amount: number;
+  lnurl: string;
+  recepientPubKey: string;
+  zappedEvent: Event;
+}): Promise<Event> => {
+  const { anonPubKey, anonPrivKey } = getAnonPubKey();
+
+  const unsignedEvent: UnsignedEvent = {
+    kind: 9734,
+    content,
+    pubkey: anonPubKey,
+    created_at: Math.floor(Date.now() / 1000),
+    tags: [
+      ["relays", "wss://relay.wavlake.com/"],
+      ["amount", amount.toString()],
+      ["lnurl", lnurl],
+      ["p", recepientPubKey],
+      ["e", zappedEvent.id],
+    ],
+  };
+  return {
+    ...unsignedEvent,
+    id: getEventHash(unsignedEvent),
+    sig: signEvent(unsignedEvent, anonPrivKey),
+  };
+};
+
+export const getLNURLFromEvent = (event: Event): string | undefined => {
+  const zapTag = event.tags.find((tag) => tag[0] === "zap");
+  if (!zapTag) return;
+
+  const [zap, zapAddress, lud] = zapTag;
+  const [username, domain] = zapAddress.split("@");
+  if (!username || !domain) return;
+
+  return `${protocol}://${domain}/.well-known/lnurlp/${username}`;
+};
+
+export const fetchLNURLInfo = async (
+  lnurl: string
+): Promise<{
+  allowsNostr?: boolean;
+  callback?: string;
+  nostrPubKey?: string;
+  error?: string;
+}> => {
+  try {
+    const res = await fetch(lnurl);
+    const {
+      allowsNostr,
+      callback,
+      maxSendable,
+      metadata,
+      minSendable,
+      nostrPubKey,
+      tag,
+    } = await res.json();
+
+    if (!validateNostrPubKey(nostrPubKey)) {
+      throw `Invalid nostr pubkey ${nostrPubKey}`;
+    }
+    if (!allowsNostr) {
+      throw "lnurl does not allow nostr";
+    }
+
+    return { allowsNostr, callback, nostrPubKey };
+  } catch (err) {
+    console.error("Error fetching lnurl info", { err, lnurl });
+    return { error: "Error fetching lnurl info" };
+  }
+};
+
+export const sendZapRequestReceivePaymentRequest = async ({
+  signedZapEvent,
+  callback,
+  milliSatAmount,
+  lnurl,
+}: {
+  signedZapEvent: Event;
+  callback: string;
+  milliSatAmount: number;
+  lnurl: string;
+}): Promise<string | undefined> => {
+  const event = JSON.stringify(signedZapEvent);
+  const paymentRequestRes = await fetch(
+    `${callback}?amount=${milliSatAmount}&nostr=${event}&lnurl=${lnurl}`
+  );
+  const { pr } = await paymentRequestRes.json();
+
+  if (!pr) {
+    console.error({ callback, lnurl, signedZapEvent });
+    throw "Error fetching payment request";
+  }
+  return pr;
 };
